Cleanup SQL stores
