--- src/Microsoft.Health.Dicom.SqlServer/Features/Store/SqlIndexDataStoreV48.cs
@@ -1,51 +1,0 @@
-// -------------------------------------------------------------------------------------------------
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
-// -------------------------------------------------------------------------------------------------
-
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Data.SqlClient;
-using Microsoft.Health.Dicom.Core.Exceptions;
-using Microsoft.Health.Dicom.SqlServer.Features.Schema;
-using Microsoft.Health.Dicom.SqlServer.Features.Schema.Model;
-using Microsoft.Health.SqlServer.Features.Client;
-
-namespace Microsoft.Health.Dicom.SqlServer.Features.Store;
-
-internal class SqlIndexDataStoreV48 : SqlIndexDataStoreV47
-{
-    public SqlIndexDataStoreV48(SqlConnectionWrapperFactory sqlConnectionWrapperFactory)
-        : base(sqlConnectionWrapperFactory)
-    {
-    }
-
-    public override SchemaVersion Version => SchemaVersion.V48;
-
-
-    public override async Task UpdateFrameDataAsync(int partitionKey, IEnumerable<long> versions, bool hasFrameMetadata, CancellationToken cancellationToken = default)
-    {
-        IEnumerable<WatermarkTableTypeRow> versionRows = versions.Select(i => new WatermarkTableTypeRow(i));
-
-        using (SqlConnectionWrapper sqlConnectionWrapper = await SqlConnectionWrapperFactory.ObtainSqlConnectionWrapperAsync(cancellationToken))
-        using (SqlCommandWrapper sqlCommandWrapper = sqlConnectionWrapper.CreateRetrySqlCommand())
-        {
-            VLatest.UpdateFrameMetadata.PopulateCommand(
-                sqlCommandWrapper,
-                partitionKey,
-                hasFrameMetadata,
-                versionRows);
-
-            try
-            {
-                await sqlCommandWrapper.ExecuteScalarAsync(cancellationToken);
-            }
-            catch (SqlException ex)
-            {
-                throw new DataStoreException(ex);
-            }
-        }
-    }
-}
