+++ test/Microsoft.Health.Dicom.Tests.Integration/Persistence/SqlDataStoreTestsFixture.cs
@@ -116,11 +116,6 @@ public class SqlDataStoreTestsFixture : IAsyncLifetime
             new[]
             {
                 new SqlIndexDataStoreV1(SqlConnectionWrapperFactory),
-                new SqlIndexDataStoreV37(SqlConnectionWrapperFactory),
-                new SqlIndexDataStoreV42(SqlConnectionWrapperFactory),
-                new SqlIndexDataStoreV44(SqlConnectionWrapperFactory),
-                new SqlIndexDataStoreV46(SqlConnectionWrapperFactory),
-                new SqlIndexDataStoreV47(SqlConnectionWrapperFactory),
                 new SqlIndexDataStoreV49(SqlConnectionWrapperFactory),
                 new SqlIndexDataStoreV50(SqlConnectionWrapperFactory),
                 new SqlIndexDataStoreV52(SqlConnectionWrapperFactory),
@@ -132,8 +127,6 @@ public class SqlDataStoreTestsFixture : IAsyncLifetime
             new[]
             {
                 new SqlInstanceStoreV1(SqlConnectionWrapperFactory),
-                new SqlInstanceStoreV34(SqlConnectionWrapperFactory),
-                new SqlInstanceStoreV46(SqlConnectionWrapperFactory),
             }));
 
         PartitionStore = new SqlPartitionStore(new VersionedCache<ISqlPartitionStore>(
@@ -174,7 +167,6 @@ public class SqlDataStoreTestsFixture : IAsyncLifetime
             new[]
             {
                 new SqlChangeFeedStoreV4(SqlConnectionWrapperFactory),
-                new SqlChangeFeedStoreV36(SqlConnectionWrapperFactory),
                 new SqlChangeFeedStoreV39(SqlConnectionWrapperFactory),
             }));
